removed unused imports and code and organized admin files in a new folder
