+++ src/components/Admin/NewProductConfirmation.tsx
